requirements.txt
astroid==1.4.4
certifi==2019.9.11
cffi==1.13.2
colorama==0.3.6
cryptography==2.8
dj-database-url==0.5.0
Django==2.0.7
django-debug-toolbar==1.4
django-heroku==0.3.1
django-querycount==0.4.1
django-split-settings==0.1.3
docopt==0.6.2
future==0.18.2
gitdb==0.6.4
GitPython==1.0.2
joe==0.0.7
lazy-object-proxy==1.2.1
numpy==1.17.3
olefile==0.46
pandas==0.25.3
pbr==1.8.1
pep8==1.7.0
Pillow==4.3.0
powerline-status==2.3
psycopg2==2.7.5
pycparser==2.19
pylint==1.5.4
pynewtonmath==1.0.1
python-dateutil==2.8.1
python-telegram-bot==11.1.0
pytz==2018.5
six==1.10.0
smmap==0.9.0
sqlparse==0.1.19
stevedore==1.10.0
tensorflow==2.0.0b1
virtualenv==14.0.5
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
whitenoise==3.3.1
wrapt==1.10.6
